New quit command. Moved commands into their own submodule of Plugins called Commands.
